+++ resource-server/src/main/java/whyarewestillalive/resourceserver/Controllers/User_Controller.java
@@ -3,6 +3,8 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -28,14 +30,18 @@ public class User_Controller {
 	}
 	
 	@GetMapping("{id}")
-	public ResponseEntity<User> getByid(@PathVariable long id){
+	public ResponseEntity<User> getByid(@PathVariable long id, @AuthenticationPrincipal Jwt jwt){
 		User user=userRepository.findById(id);
-		if(user==null) {
-			return ResponseEntity.notFound().build();
+		String authorities=jwt.getClaims().get("authorities").toString();
+
+		if (authorities.equals("ROLE_ADMIN")) {
+			if (user == null) {
+				return ResponseEntity.notFound().build();
+			} else {
+				return ResponseEntity.ok(user);
+			}
 		}
-		else {
-			return ResponseEntity.ok(user);
-		}
+		return ResponseEntity.badRequest().build();
 	}
 	
 	@PostMapping
@@ -51,28 +57,42 @@ public class User_Controller {
 	}
 	
 	@PutMapping
-	public ResponseEntity<?> update(@RequestBody User user){
+	public ResponseEntity<?> update(@RequestBody User user, @AuthenticationPrincipal Jwt jwt){
 		User temp=userRepository.findById(user.getId());
-		if(temp==null) {
-			return ResponseEntity.notFound().build();
+		String authorities=jwt.getClaims().get("authorities").toString();
+		String name=jwt.getClaims().get("client_id").toString();
+
+		if ((user.getName().equals(name)) || (authorities.equals("ROLE_ADMIN"))) {
+
+			if (temp == null) {
+				return ResponseEntity.notFound().build();
+			} else {
+				temp = userRepository.save(user);
+				return ResponseEntity.ok().build();
+			}
 		}
-		else {
-			temp=userRepository.save(user);
-			return ResponseEntity.ok().build();
-		}
+
+		return ResponseEntity.badRequest().build();
 	}
 	
 	
 	@DeleteMapping
-	public ResponseEntity<?> delete(@RequestBody long id){
+	public ResponseEntity<?> delete(@RequestBody long id, @AuthenticationPrincipal Jwt jwt){
 		User user=userRepository.findById(id);
-		if(user==null) {
-			return ResponseEntity.notFound().build();
+		String authorities=jwt.getClaims().get("authorities").toString();
+		String name=jwt.getClaims().get("client_id").toString();
+
+		if((user.getName().equals(name)) || (authorities.equals("ROLE_ADMIN"))) {
+
+			if (user == null) {
+				return ResponseEntity.notFound().build();
+			} else {
+				userRepository.deleteById(id);
+				return ResponseEntity.ok().build();
+			}
 		}
-		else {
-			userRepository.deleteById(id);
-			return ResponseEntity.ok().build();
-		}
+
+		return ResponseEntity.badRequest().build();
 	}
 	
 }
